requirements.txt
selenium==4.1.3
fake-useragent==0.1.11
sqlalchemy==1.4.36
pandas==1.4.2
python-dotenv==0.20.0
loguru==0.6.0
schedule==1.1.0
tqdm==4.64.0